Keep lua script in a .lua file
